+++ generators/cleanup.js
@@ -487,13 +487,6 @@ function cleanupOldServerFiles(generator, javaDir, testDir, mainResourceDir, tes
   if (generator.isJhipsterVersionLessThan('7.0.0-beta.1')) {
     generator.removeFile(`${javaDir}config/CloudDatabaseConfiguration.java`);
   }
-  if (generator.isJhipsterVersionLessThan('7.1.1')) {
-    generator.removeFile(`${javaDir}repository/ReactiveN1qlCouchbaseRepository.java `);
-    generator.removeFile(`${testDir}config/DatabaseConfigurationIT.java`);
-    if (generator.searchEngine !== 'couchbase') {
-      generator.removeFile(`${javaDir}repository/CustomN1qlCouchbaseRepository.java`);
-    }
-  }
 }
 
 /**
